wip backend tests for python job and deno flow
Worker modifed to drop the `broadcast::Sender` after subscribing a
`broadcast::Receiver` to it.  If the last Sender is closed, the
Recievers are closed.  But this will never happen unless run_worker
drops the Sender passed to it.

Tests don't go through the web/http stuff so they skip serialization and
all that. Not sure if this is good or bad, maybe we want both.

each test marked with `#[sqlx::test]` is run in a new database named
_sqlx_test_N. There's quite a bit of overhead to this compared to
running tests in transactions but it provides more isolation and the
implmentation is a bit simpler.

backend/src/fixtures/base.sql is run by `#[sqlx::test]` to set up some
data when testing.  It also adds a trigger used to synchronize a bit,
it's kind of a hack.  Ideally the testing database scheam doesn't differ
from the actual schema.