allow GET to resume job
I way overthought this. I figured we'd just deserialize all items in the
query string as a serde_json::Value and assumed it would do a kind of
best-effort to detect numbers, booleans, and stuff like that.  But it
doesn't do that at all, it always deserializes values as strings.

So
    GET /resume/job?something=123&foobar=false
will resume the job with
    {"something": "123", "foobar": "false"}
Which is a bit silly.  But I'm not sure of a good alternative
to support more types that is less silly than URL quoting a JSON
document like %7B%22something%22%3A123%2C%22foobar%22%3Afalse%7D